requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.27.2
anthropic==0.39.0
google-api-python-client==2.154.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
supabase==2.10.0
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.1
python-dateutil==2.9.0
pytz==2024.2
jinja2==3.1.4
python-multipart==0.0.18
